# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/api/v1/order/create" : {
      "post" : {
        "summary" : "创建订单",
        "description" : "创建订单",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "userId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "couponId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : "状态(1-已创建,2-已使用,3-已过期,4-已取消)"
                  },
                  "getTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "expireTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "useTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "cancelTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "createdByJava" : {
                    "type" : "integer",
                    "description" : "来源标识字段",
                    "default" : 0
                  },
                  "createdByGo" : {
                    "type" : "integer",
                    "description" : "",
                    "default" : 0
                  },
                  "requestId" : {
                    "type" : "string",
                    "description" : "新增字段"
                  },
                  "version" : {
                    "type" : "integer",
                    "description" : "固定为0（Java端创建不影响Go端标识）"
                  }
                },
                "description" : "订单信息"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "userId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "couponId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : "状态(1-已创建,2-已使用,3-已过期,4-已取消)"
                        },
                        "getTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "expireTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "useTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "cancelTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "createdByJava" : {
                          "type" : "integer",
                          "description" : "来源标识字段",
                          "default" : 0
                        },
                        "createdByGo" : {
                          "type" : "integer",
                          "description" : "",
                          "default" : 0
                        },
                        "requestId" : {
                          "type" : "string",
                          "description" : "新增字段"
                        },
                        "version" : {
                          "type" : "integer",
                          "description" : "固定为0（Java端创建不影响Go端标识）"
                        }
                      },
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/seckill/create" : {
      "post" : {
        "summary" : "创建秒杀订单",
        "description" : "创建秒杀订单",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "description" : "用户ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "couponId",
          "in" : "query",
          "description" : "优惠券ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "userId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "couponId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : "状态(1-已创建,2-已使用,3-已过期,4-已取消)"
                        },
                        "getTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "expireTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "useTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "cancelTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "createdByJava" : {
                          "type" : "integer",
                          "description" : "来源标识字段",
                          "default" : 0
                        },
                        "createdByGo" : {
                          "type" : "integer",
                          "description" : "",
                          "default" : 0
                        },
                        "requestId" : {
                          "type" : "string",
                          "description" : "新增字段"
                        },
                        "version" : {
                          "type" : "integer",
                          "description" : "固定为0（Java端创建不影响Go端标识）"
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/seckill/cancel" : {
      "post" : {
        "summary" : "取消秒杀订单",
        "description" : "取消秒杀订单",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "query",
          "description" : "订单ID",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "description" : "用户ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "boolean",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/seckill/check/{userId}/{couponId}" : {
      "get" : {
        "summary" : "检查用户是否已领取优惠券",
        "description" : "检查用户是否已领取优惠券",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "用户ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "couponId",
          "in" : "path",
          "description" : "优惠券ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "boolean",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/seckill/user/{userId}" : {
      "get" : {
        "summary" : "查询用户订单",
        "description" : "查询用户订单",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "用户ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "pageNum",
          "in" : "query",
          "description" : "页码",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "1"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "每页数量",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "10"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "array",
                      "description" : "",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "userId" : {
                            "type" : "integer",
                            "description" : "",
                            "format" : "int64"
                          },
                          "couponId" : {
                            "type" : "integer",
                            "description" : "",
                            "format" : "int64"
                          },
                          "status" : {
                            "type" : "integer",
                            "description" : "状态(1-已创建,2-已使用,3-已过期,4-已取消)"
                          },
                          "getTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "expireTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "useTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "cancelTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "createTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "updateTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "createdByJava" : {
                            "type" : "integer",
                            "description" : "来源标识字段",
                            "default" : 0
                          },
                          "createdByGo" : {
                            "type" : "integer",
                            "description" : "",
                            "default" : 0
                          },
                          "requestId" : {
                            "type" : "string",
                            "description" : "新增字段"
                          },
                          "version" : {
                            "type" : "integer",
                            "description" : "固定为0（Java端创建不影响Go端标识）"
                          }
                        },
                        "description" : "com.aliyun.seckill.common.pojo.Order"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/seckill/admin" : {
      "get" : {
        "summary" : "管理员查询所有订单",
        "description" : "管理员查询所有订单",
        "parameters" : [ {
          "name" : "pageNum",
          "in" : "query",
          "description" : "页码",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "1"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "每页数量",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "10"
          }
        }, {
          "name" : "startTime",
          "in" : "query",
          "description" : "开始时间",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "endTime",
          "in" : "query",
          "description" : "结束时间",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "array",
                      "description" : "",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "userId" : {
                            "type" : "integer",
                            "description" : "",
                            "format" : "int64"
                          },
                          "couponId" : {
                            "type" : "integer",
                            "description" : "",
                            "format" : "int64"
                          },
                          "status" : {
                            "type" : "integer",
                            "description" : "状态(1-已创建,2-已使用,3-已过期,4-已取消)"
                          },
                          "getTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "expireTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "useTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "cancelTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "createTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "updateTime" : {
                            "type" : "string",
                            "description" : ""
                          },
                          "createdByJava" : {
                            "type" : "integer",
                            "description" : "来源标识字段",
                            "default" : 0
                          },
                          "createdByGo" : {
                            "type" : "integer",
                            "description" : "",
                            "default" : 0
                          },
                          "requestId" : {
                            "type" : "string",
                            "description" : "新增字段"
                          },
                          "version" : {
                            "type" : "integer",
                            "description" : "固定为0（Java端创建不影响Go端标识）"
                          }
                        },
                        "description" : "com.aliyun.seckill.common.pojo.Order"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}